backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
